@@ -5,22 +5,22 @@
 	<parent>
 		<groupId>org.springframework.boot</groupId>
 		<artifactId>spring-boot-starter-parent</artifactId>
-		<version>2.7.6</version>
+		<version>3.2.2</version>
 		<relativePath/> <!-- lookup parent from repository -->
 	</parent>
 	<groupId>de.skotstein.lib.spring</groupId>
 	<artifactId>restfulspring</artifactId>
-	<version>0.0.2-SNAPSHOT</version>
+	<version>3.0.2-SNAPSHOT</version>
 	<name>restfulspring</name>
 	<description>Demo project for Spring Boot</description>
 	<properties>
-		<java.version>13</java.version>
+		<java.version>17</java.version>
 	</properties>
 	<dependencies>
-		<dependency>
+		<!--<dependency>
 			<groupId>org.springframework.boot</groupId>
 			<artifactId>spring-boot-starter</artifactId>
-		</dependency>
+		</dependency>-->
 
 		<dependency>
 			<groupId>org.springframework.boot</groupId>
@@ -31,7 +31,6 @@
 		<dependency>
     		<groupId>com.fasterxml.jackson.dataformat</groupId>
     		<artifactId>jackson-dataformat-xml</artifactId>
-    		<version>2.11.1</version>
 		</dependency>
 
 		<!-- JPA -->
