@@ -20,7 +20,8 @@ package de.skotstein.lib.spring.restfulspring.config;
 import java.text.SimpleDateFormat;
 import java.util.Date;
 
-import javax.servlet.http.HttpServletRequest;
+//import javax.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletRequest;
 
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.MediaType;
@@ -29,20 +30,21 @@ import org.springframework.web.server.ResponseStatusException;
 
 import de.skotstein.lib.spring.restfulspring.model.entities.Error;
 
+
 public abstract class DefaultErrorBuilder extends AbstractErrorBuilder{
 
     @Override
     public ResponseEntity<String> build(ResponseStatusException ex, HttpServletRequest request, HttpHeaders httpHeaders) {
         httpHeaders.add("Content-Type", getDefaultMediaType().toString());
-        return new ResponseEntity<String>(serializeError(createError(ex, request)),httpHeaders,ex.getStatus());
+        return new ResponseEntity<String>(serializeError(createError(ex, request)),httpHeaders,ex.getStatusCode());
     }
 
     protected Error createError(ResponseStatusException ex, HttpServletRequest request){
         Error error = new Error();
         SimpleDateFormat formatter = new SimpleDateFormat("dd-MM-yyyy HH:mm:ss");
         error.setTimestamp(formatter.format(new Date()));
-        error.setStatus(ex.getStatus().value());
-        error.setError(ex.getStatus().getReasonPhrase());
+        error.setStatus(ex.getStatusCode().value());
+        error.setError(ex.getStatusCode().toString());
         error.setMessage(ex.getReason());
         error.setPath(request.getRequestURI());
         error.setRequest(request);
